+++ all-is-cubes/src/math.rs
@@ -1,7 +1,5 @@
 //! Mathematical utilities and decisions.
 
-use std::fmt;
-
 use cgmath::{Point3, Vector3};
 use num_traits::identities::Zero;
 pub use ordered_float::{FloatIsNan, NotNan};
@@ -10,6 +8,8 @@ use crate::util::{ConciseDebug, CustomFormat};
 
 mod aab;
 pub use aab::*;
+mod axis;
+pub use axis::*;
 #[macro_use]
 mod color;
 pub use color::*;